refactor amenity seed data to another file
