+++ database/database.js
@@ -2,24 +2,10 @@ require('dotenv').config();
 
 const mongoose = require('mongoose');
 
-console.log();
-
 mongoose.connect(process.env.URL, { useNewUrlParser: true });
 const db = mongoose.connection;
 
 db.on('error', console.error.bind(console, 'connection error:'));
-
-const descriptionsSchema = new mongoose.Schema({
-  id: Number,
-  name: String,
-  ownerName: String,
-  location: String,
-  livingSpace: String,
-  beds: Number,
-  baths: Number,
-  description: String,
-  homehighlights: {},
-});
 
 const amenitiesSchema = new mongoose.Schema({
   id: Number,
@@ -31,7 +17,20 @@ const amenitiesSchema = new mongoose.Schema({
   notIncluded: [String],
 });
 
-const Description = mongoose.model('descriptions', descriptionsSchema);
-const Amenity = mongoose.model('amenities', amenitiesSchema);
+const descriptionsSchema = new mongoose.Schema({
+  id: Number,
+  name: String,
+  ownerName: String,
+  location: String,
+  livingSpace: String,
+  beds: Number,
+  baths: Number,
+  description: String,
+  homehighlights: {},
+  amenities: amenitiesSchema,
+});
 
-module.exports = { Description, Amenity };
+const Description = mongoose.model('descriptions', descriptionsSchema);
+// const Amenity = mongoose.model('amenities', amenitiesSchema);
+
+module.exports = { Description };
