+++ package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "dotenv": "^6.0.0",
     "express": "^4.16.3",
+    "faker": "^4.1.0",
     "mongodb": "^3.1.4",
     "mongoose": "^5.2.13"
   },
@@ -28,7 +29,6 @@
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.1",
     "eslint-plugin-react": "^7.11.1",
-    "faker": "^4.1.0",
     "webpack": "^4.17.2"
   }
 }
